cleanup trap code

+++ bld/trap/lcl/os2v2/c/os2v2acc.c
@@ -103,7 +103,7 @@ static ULONG            ExceptLinear;
 static UCHAR            TypeProcess;
 static BOOL             Is32Bit;
 static watch_point      WatchPoints[MAX_WATCHES];
-static short            WatchCount = 0;
+static int              WatchCount = 0;
 static unsigned_16      lastCS;
 static unsigned_16      lastSS;
 static unsigned_32      lastEIP;
@@ -1288,7 +1288,7 @@ trap_retval TRAP_CORE( Set_watch )( void )
     acc = GetInPtr( 0 );
     ret = GetOutPtr( 0 );
     ret->multiplier = 50000;
-    ret->err = 1;       // fail
+    ret->err = 1;       // failure
     if( WatchCount < MAX_WATCHES ) { // nyi - artificial limit (32 should be lots)
         ret->err = 0;   // OK
         wp = WatchPoints + WatchCount;
